requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
crypto==1.4.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
Naked==0.1.31
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
sqlparse==0.4.2
urllib3==1.26.7
